+++ src/components/Auth.tsx
@@ -9,6 +9,8 @@ import ClientOnly from "./ClientOnly";
 import { H2 } from "./Text";
 import Select from "./Select";
 import FormLabel from "./form/FormLabel";
+import { useActions } from "opds-web-client/lib/components/context/ActionsContext";
+import useTypedSelector from "hooks/useTypedSelector";
 
 /**
  *  - makes sure auth state is loaded from cookies
@@ -20,6 +22,23 @@ const Auth: React.FC = ({ children }) => {
   const dialog = useDialogState();
   const library = useLibraryContext();
   const [authProvider, setAuthProvider] = React.useState(providers?.[0]);
+  const { fetcher, actions, dispatch } = useActions();
+
+  /**
+   * This component is responsible for fetching loans whenever the
+   * auth credentials change
+   */
+  const { provider: currentProvider, credentials } =
+    fetcher.getAuthCredentials() ?? {};
+  const loansUrl = useTypedSelector(state => state.loans.url);
+  React.useEffect(() => {
+    if (currentProvider && credentials) {
+      dispatch(
+        actions.saveAuthCredentials({ provider: currentProvider, credentials })
+      );
+      if (loansUrl) dispatch(actions.fetchLoans(loansUrl));
+    }
+  }, [currentProvider, credentials, loansUrl, actions, dispatch]);
 
   // the providers get set when the form is shown, so
   // it's initially undefined. We need to update when they get set